+++ app/src/main/java/com/example/releasethekraken/MainActivity.java
@@ -1,0 +1,15 @@
+package com.example.releasethekraken;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        setContentView(R.layout.activity_main);
+    }
+}
