+++ app/src/main/java/com/example/releasethekraken/view/ViewProfileFragment.java
@@ -1,4 +1,4 @@
-package com.example.releasethekraken;
+package com.example.releasethekraken.view;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -11,6 +11,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 
+import com.example.releasethekraken.R;
 import com.example.releasethekraken.databinding.FragmentViewProfileBinding;
 
 public class ViewProfileFragment extends Fragment {
