+++ app/src/main/res/layout/fragment_account_create.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/fragment_vertical_margin"
     android:paddingRight="@dimen/fragment_horizontal_margin"
     android:paddingBottom="@dimen/fragment_vertical_margin"
-    tools:context=".ui.login.AccountCreateFragment">
+    tools:context=".view.ui.login.AccountCreateFragment">
 
     <TextView
         android:id="@+id/account_creation_welcome"
