+++ app/src/main/res/layout/fragment_login.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/fragment_vertical_margin"
     android:paddingRight="@dimen/fragment_horizontal_margin"
     android:paddingBottom="@dimen/fragment_vertical_margin"
-    tools:context=".ui.login.LoginFragment">
+    tools:context=".view.ui.login.LoginFragment">
 
     <TextView
         android:id="@+id/login_welcome"
