+++ app/src/main/res/layout/fragment_main_menu.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/black"
     android:theme="@style/ThemeOverlay.ReleaseTheKraken.FullscreenContainer"
-    tools:context=".MainMenuFragment">
+    tools:context=".view.MainMenuFragment">
 
     <!-- The primary full-screen view. This can be replaced with whatever view
          is needed to present your content, e.g. VideoView, SurfaceView,
